Added a figure and caption to section 2A.
Also, added placeholder figures for the figures we may want to include. The figure added is composed of the placeholder figures. Replacing the placeholders in the repository will automatically replace them in the paper.

I have gone well beyond the end of my latex knowledge to create the figure, so it may not be done in the best way.