Added footnotes for follow up questions.
I still highly recommend that we do not go into a lot of detail about the way we refined our question because I think this is a tangential point and the paper will lose focus. Instead, I would argue that we comment in a footnote (similar to how we address the torsional wave) that center of mass is another interesting avenue to attack this problem which we do not have space to discuss in the present work. I have added such a footnote.

Also fleshed out the footnote for the torsional wave.

Re-commented out the table of slinky parameters as we discussed on Saturday. Modifications have been correctly merged.

Fixed a problem where I think Jesse was supposed to be assigned to the multi-frame figure instead of the center of mass comment.